+++ packages/data-mate/src/column/index.ts
@@ -1,3 +1,5 @@
 export * from './column';
+export * from './interfaces';
 export * from './utils';
 export * from './transforms';
+export * from './validators';
