+++ packages/data-mate/test/column/column-number-spec.ts
@@ -1,72 +1,10 @@
 import 'jest-fixtures';
 import { FieldType, Maybe } from '@terascope/types';
 import {
-    BigIntVector,
-    Column, ColumnTransform, Vector
-} from '../src';
-import { ColumnValidator } from '../src/column/validators';
+    BigIntVector, Column, Vector
+} from '../../src';
 
-describe('Column', () => {
-    describe(`when field type is ${FieldType.Keyword}`, () => {
-        let col: Column<string>;
-        const values: Maybe<string>[] = [
-            'Batman',
-            'Robin',
-            'Superman',
-            null,
-            'SpiderMan',
-        ];
-        beforeEach(() => {
-            col = Column.fromJSON<string>({
-                name: 'name',
-                config: {
-                    type: FieldType.Keyword,
-                },
-            }, values);
-        });
-
-        it('should have the correct size', () => {
-            expect(col.count()).toEqual(values.length);
-        });
-
-        it('should have the same id when forked with the same vector', () => {
-            expect(col.fork().id).toEqual(col.id);
-        });
-
-        it('should NOT have the same id when forked with a different vector', () => {
-            const vector = col.vector.slice(0, 2);
-            expect(col.fork(vector).id).not.toEqual(col.id);
-        });
-
-        it('should be able to iterate over the values', () => {
-            expect([...col]).toEqual(values);
-            expect(col.toJSON()).toEqual(values);
-        });
-
-        it('should be able to get the Vector', () => {
-            expect(col.vector).toBeInstanceOf(Vector);
-        });
-
-        it('should be able to validate the values', () => {
-            const newCol = col.validate(ColumnValidator.isURL);
-            expect(newCol.id).not.toBe(col.id);
-            expect([...newCol]).toEqual(values.map(() => null));
-        });
-
-        it('should be able to transform the column using toUpperCase', () => {
-            const newCol = col.transform(ColumnTransform.toUpperCase);
-
-            expect(newCol.id).not.toBe(col.id);
-            expect(newCol.config).toEqual(col.config);
-            expect([...newCol]).toEqual(values.map((value) => {
-                if (typeof value === 'string') return value.toUpperCase();
-                return null;
-            }));
-        });
-
-        test.todo('should be immutable');
-    });
-
+describe('Column (Number Types)', () => {
     describe(`when field type is ${FieldType.Short}`, () => {
         let col: Column<number>;
         const values: Maybe<number>[] = [
@@ -76,6 +14,7 @@ describe('Column', () => {
             null,
             4,
         ];
+
         beforeEach(() => {
             col = Column.fromJSON<number>({
                 name: 'score',
