+++ packages/data-mate/test/column/column-string-spec.ts
@@ -1,0 +1,68 @@
+import 'jest-fixtures';
+import { FieldType, Maybe } from '@terascope/types';
+import {
+    ColumnValidator,
+    Column, ColumnTransform, Vector
+} from '../../src';
+
+describe('Column (Number Types)', () => {
+    describe(`when field type is ${FieldType.Keyword}`, () => {
+        let col: Column<string>;
+        const values: Maybe<string>[] = [
+            'Batman',
+            'Robin',
+            'Superman',
+            null,
+            'SpiderMan',
+        ];
+        beforeEach(() => {
+            col = Column.fromJSON<string>({
+                name: 'name',
+                config: {
+                    type: FieldType.Keyword,
+                },
+            }, values);
+        });
+
+        it('should have the correct size', () => {
+            expect(col.count()).toEqual(values.length);
+        });
+
+        it('should have the same id when forked with the same vector', () => {
+            expect(col.fork().id).toEqual(col.id);
+        });
+
+        it('should NOT have the same id when forked with a different vector', () => {
+            const vector = col.vector.slice(0, 2);
+            expect(col.fork(vector).id).not.toEqual(col.id);
+        });
+
+        it('should be able to iterate over the values', () => {
+            expect([...col]).toEqual(values);
+            expect(col.toJSON()).toEqual(values);
+        });
+
+        it('should be able to get the Vector', () => {
+            expect(col.vector).toBeInstanceOf(Vector);
+        });
+
+        it('should be able to validate the values', () => {
+            const newCol = col.validate(ColumnValidator.isURL);
+            expect(newCol.id).not.toBe(col.id);
+            expect([...newCol]).toEqual(values.map(() => null));
+        });
+
+        it('should be able to transform the column using toUpperCase', () => {
+            const newCol = col.transform(ColumnTransform.toUpperCase);
+
+            expect(newCol.id).not.toBe(col.id);
+            expect(newCol.config).toEqual(col.config);
+            expect([...newCol]).toEqual(values.map((value) => {
+                if (typeof value === 'string') return value.toUpperCase();
+                return null;
+            }));
+        });
+
+        test.todo('should be immutable');
+    });
+});
